+++ cmn-data-svc-server/src/main/java/org/acumos/cds/util/ApiPageable.java
@@ -1,0 +1,47 @@
+/*-
+ * ===============LICENSE_START=======================================================
+ * Acumos
+ * ===================================================================================
+ * Copyright (C) 2017 AT&T Intellectual Property & Tech Mahindra. All rights reserved.
+ * ===================================================================================
+ * This Acumos software file is distributed by AT&T and Tech Mahindra
+ * under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *  
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *  
+ * This file is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ===============LICENSE_END=========================================================
+ */
+
+package org.acumos.cds.util;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+
+/**
+ * Annotation used to decorate REST endpoints that accept a Spring Pageable
+ * request parameter so Swagger shows the appropriate details.
+ * 
+ * https://stackoverflow.com/questions/35404329/swagger-documentation-for-spring-pageable-interface
+ */
+@Target({ ElementType.METHOD, ElementType.ANNOTATION_TYPE, ElementType.TYPE })
+@Retention(RetentionPolicy.RUNTIME)
+@ApiImplicitParams({
+		@ApiImplicitParam(name = "page", dataType = "int", paramType = "query", value = "Results page you want to retrieve (0..N)"),
+		@ApiImplicitParam(name = "size", dataType = "int", paramType = "query", value = "Number of records per page."),
+		@ApiImplicitParam(name = "sort", allowMultiple = true, dataType = "string", paramType = "query", //
+				value = "Sorting criteria in the format: property(,asc|desc). " //
+						+ "Default sort order is ascending. " //
+						+ "Multiple sort criteria are supported.") })
+public @interface ApiPageable {
+}